requirements.txt
pypdf==4.0.1
python-dotenv==1.0.0